Handle relative redirect locations in header when fetching RSS feeds
